+++ src/bezier.js
@@ -1,0 +1,274 @@
+// This code was extracted from Bezier.js
+// A javascript Bezier curve library by Pomax.
+// Based on http://pomax.github.io/bezierinfo
+// http://pomax.github.io/bezierjs/
+
+(function() {
+  "use strict";
+
+  // Math functions. I hate the Math namespace with a passion.
+  var abs = Math.abs,
+      min = Math.min,
+      max = Math.max,
+      cos = Math.cos,
+      sin = Math.sin,
+      acos = Math.acos,
+      asin = Math.asin,
+      atan2 = Math.atan2,
+      sqrt = Math.sqrt,
+      // cube root function yielding real roots
+      crt = function(v) { if(v<0) return -Math.pow(-v,1/3); return Math.pow(v,1/3); },
+      pi = Math.PI,
+      tau = 2*pi,
+      quart = pi/2;
+
+  // a zero coordinate, which is surprisingly useful
+  var ZERO = {x:0,y:0,z:0};
+
+  // Bezier utility functions
+  var utils = {
+    // Legendre-Gauss abscissae with n=24 (x_i values, defined at i=n as the roots of the nth order Legendre polynomial Pn(x))
+    Tvalues: [
+      -0.0640568928626056260850430826247450385909,
+       0.0640568928626056260850430826247450385909,
+      -0.1911188674736163091586398207570696318404,
+       0.1911188674736163091586398207570696318404,
+      -0.3150426796961633743867932913198102407864,
+       0.3150426796961633743867932913198102407864,
+      -0.4337935076260451384870842319133497124524,
+       0.4337935076260451384870842319133497124524,
+      -0.5454214713888395356583756172183723700107,
+       0.5454214713888395356583756172183723700107,
+      -0.6480936519369755692524957869107476266696,
+       0.6480936519369755692524957869107476266696,
+      -0.7401241915785543642438281030999784255232,
+       0.7401241915785543642438281030999784255232,
+      -0.8200019859739029219539498726697452080761,
+       0.8200019859739029219539498726697452080761,
+      -0.8864155270044010342131543419821967550873,
+       0.8864155270044010342131543419821967550873,
+      -0.9382745520027327585236490017087214496548,
+       0.9382745520027327585236490017087214496548,
+      -0.9747285559713094981983919930081690617411,
+       0.9747285559713094981983919930081690617411,
+      -0.9951872199970213601799974097007368118745,
+       0.9951872199970213601799974097007368118745
+    ],
+
+    // Legendre-Gauss weights with n=24 (w_i values, defined by a function linked to in the Bezier primer article)
+    Cvalues: [
+      0.1279381953467521569740561652246953718517,
+      0.1279381953467521569740561652246953718517,
+      0.1258374563468282961213753825111836887264,
+      0.1258374563468282961213753825111836887264,
+      0.1216704729278033912044631534762624256070,
+      0.1216704729278033912044631534762624256070,
+      0.1155056680537256013533444839067835598622,
+      0.1155056680537256013533444839067835598622,
+      0.1074442701159656347825773424466062227946,
+      0.1074442701159656347825773424466062227946,
+      0.0976186521041138882698806644642471544279,
+      0.0976186521041138882698806644642471544279,
+      0.0861901615319532759171852029837426671850,
+      0.0861901615319532759171852029837426671850,
+      0.0733464814110803057340336152531165181193,
+      0.0733464814110803057340336152531165181193,
+      0.0592985849154367807463677585001085845412,
+      0.0592985849154367807463677585001085845412,
+      0.0442774388174198061686027482113382288593,
+      0.0442774388174198061686027482113382288593,
+      0.0285313886289336631813078159518782864491,
+      0.0285313886289336631813078159518782864491,
+      0.0123412297999871995468056670700372915759,
+      0.0123412297999871995468056670700372915759
+    ],
+    arcfn: function(t, derivativeFn) {
+      var d = derivativeFn(t);
+      var l = d.x*d.x + d.y*d.y;
+      if(typeof d.z !== "undefined") {
+        l += d.z*d.z;
+      }
+      return sqrt(l);
+    },
+    length: function(derivativeFn) {
+      var z=0.5,sum=0,len=this.Tvalues.length,i,t;
+      for(i=0; i<len; i++) {
+        t = z * this.Tvalues[i] + z;
+        sum += this.Cvalues[i] * this.arcfn(t,derivativeFn);
+      }
+      return z * sum;
+    },
+    angle: function(o,v1,v2) {
+      var dx1 = v1.x - o.x,
+          dy1 = v1.y - o.y,
+          dx2 = v2.x - o.x,
+          dy2 = v2.y - o.y,
+          cross = dx1*dy2 - dy1*dx2,
+          m1 = sqrt(dx1*dx1+dy1*dy1),
+          m2 = sqrt(dx2*dx2+dy2*dy2),
+          dot;
+      dx1/=m1; dy1/=m1; dx2/=m2; dy2/=m2;
+      dot = dx1*dx2 + dy1*dy2;
+      return atan2(cross, dot);
+    },
+    dist: function(p1, p2) {
+      var dx = p1.x - p2.x,
+          dy = p1.y - p2.y;
+      return sqrt(dx*dx+dy*dy);
+    },
+    align: function(points, line) {
+      var tx = line.p1.x,
+          ty = line.p1.y,
+          a = -atan2(line.p2.y-ty, line.p2.x-tx),
+          d = function(v) {
+            return {
+              x: (v.x-tx)*cos(a) - (v.y-ty)*sin(a),
+              y: (v.x-tx)*sin(a) + (v.y-ty)*cos(a)
+            };
+          };
+      return points.map(d);
+    }
+  };
+
+  /**
+   * Bezier curve constructor. The constructor argument can be one of three things:
+   *
+   * 1. array/4 of {x:..., y:..., z:...}, z optional
+   * 2. numerical array/8 ordered x1,y1,x2,y2,x3,y3,x4,y4
+   * 3. numerical array/12 ordered x1,y1,z1,x2,y2,z2,x3,y3,z3,x4,y4,z4
+   *
+   */
+  var Bezier = function(coords) {
+    var args = (coords && coords.forEach ? coords : arguments);
+    if(typeof args[0] === "object") {
+      args = [];
+      for(var i=0; i<coords.length; i++) {
+        ['x','y','z'].forEach(function(d) {
+          if(typeof coords[i][d] !== "undefined") {
+            args.push(coords[i][d]);
+          }
+        });
+      }
+    }
+    var len = args.length;
+    if(len!==6 && len!==8 && len!==9 && len!==12) {
+      console.log(coords);
+      throw new Error("This Bezier curve library only supports quadratic and cubic curves (in 2d and 3d)");
+    }
+    var _3d = (len === 9 || len === 12);
+    this._3d = _3d;
+    var points = [];
+    for(var idx=0, step=(_3d ? 3 : 2); idx<len; idx+=step) {
+      var point = {
+        x: args[idx],
+        y: args[idx+1]
+      };
+      if(_3d) { point.z = args[idx+2] };
+      points.push(point);
+    }
+    this.order = points.length - 1;
+    this.points = points;
+    var dims = ['x','y'];
+    if(_3d) dims.push('z');
+    this.dims = dims;
+    this.dimlen = dims.length;
+    (function(curve) {
+      var a = utils.align(points, {p1:points[0], p2:points[curve.order]});
+      for(var i=0; i<a.length; i++) {
+        if(abs(a[i].y) > 0.0001) {
+          curve._linear = false;
+          return;
+        }
+      }
+      curve._linear = true;
+    }(this));
+    this._t1 = 0;
+    this._t2 = 1;
+    this.update();
+  };
+
+  Bezier.utils = utils;
+
+  Bezier.prototype = {
+    update: function() {
+      // one-time compute derivative coordinates
+      this.dpoints = [];
+      for(var p=this.points, d=p.length, c=d-1; d>1; d--, c--) {
+        var list = [];
+        for(var j=0, dpt; j<c; j++) {
+          dpt = {
+            x: c * (p[j+1].x - p[j].x),
+            y: c * (p[j+1].y - p[j].y)
+          };
+          if(this._3d) {
+            dpt.z = c * (p[j+1].z - p[j].z);
+          }
+          list.push(dpt);
+        }
+        this.dpoints.push(list);
+        p = list;
+      };
+      this.computedirection();
+    },
+    computedirection: function() {
+      var points = this.points;
+      var angle = utils.angle(points[0], points[this.order], points[1]);
+      this.clockwise = angle > 0;
+    },
+    length: function() {
+      return utils.length(this.derivative.bind(this));
+    },
+    get: function(t) {
+      return this.compute(t);
+    },
+    compute: function(t) {
+      // shortcuts
+      if(t===0) { return this.points[0]; }
+      if(t===1) { return this.points[this.order]; }
+      // plain computation
+      var mt = 1-t,
+          mt2 = mt*mt,
+          t2 = t*t,
+          a,b,c,d = 0,
+          p = this.points;
+      if(this.order===2) {
+        p = [p[0], p[1], p[2], ZERO];
+        a = mt2;
+        b = mt*t*2;
+        c = t2;
+      }
+      if(this.order===3) {
+        a = mt2*mt;
+        b = mt2*t*3;
+        c = mt*t2*3;
+        d = t*t2;
+      }
+      var ret = {
+        x: a*p[0].x + b*p[1].x + c*p[2].x + d*p[3].x,
+        y: a*p[0].y + b*p[1].y + c*p[2].y + d*p[3].y
+      };
+      if(this._3d) {
+        ret.z = a*p[0].z + b*p[1].z + c*p[2].z + d*p[3].z;
+      }
+      return ret;
+    },
+    derivative: function(t) {
+      var mt = 1-t,
+          a,b,c=0,
+          p = this.dpoints[0];
+      if(this.order===2) { p = [p[0], p[1], ZERO]; a = mt; b = t; }
+      if(this.order===3) { a = mt*mt; b = mt*t*2; c = t*t; }
+      var ret = {
+        x: a*p[0].x + b*p[1].x + c*p[2].x,
+        y: a*p[0].y + b*p[1].y + c*p[2].y
+      };
+      if(this._3d) {
+        ret.z = a*p[0].z + b*p[1].z + c*p[2].z;
+      }
+      return ret;
+    }
+  };
+
+  module.exports = Bezier;
+
+}());